pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "flask>=3.1.1",
    "python-telegram-bot>=20.0",
]
